+++ HshpGame_GD/StageObject/ObjectJumpPad.cpp
@@ -23,5 +23,5 @@ void ObjectJumpPad::Update()
 
 void ObjectJumpPad::Draw()
 {
-	DrawCircle(GetCenterX(), GetCenterY() + Game::kBlockSize - (Game::kBlockSize / 4), Game::kBlockSize / 2 - 2, 0xffff00, true);
+	DrawCircle(static_cast<int>(GetCenterX()), static_cast<int>(GetCenterY() + Game::kBlockSize - (Game::kBlockSize / 4)), static_cast<int>(Game::kBlockSize / 2 - 2), 0xffff00, true);
 }
